requirements.txt
pdf2image
Pillow 
transformers
safetensors
torch==2.3.0
torchvision==0.18.0
pytorch-lightning==2.3.0
huggingface_hub
einops
timm
requests
opencv-python
numpy
regex
pubchempy
openai
huggingface_hub
streamlit
python-dotenv
uvicorn
fastapi
python-multipart